financial-analyst-agent/requirements.txt
# Core dependencies
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
python-multipart>=0.0.9

# File processing
PyPDF2>=3.0.0
Pillow>=10.0.0
pandas>=2.0.0

# MCP (Model Context Protocol) support
mcp>=1.1.0

# Utilities
python-dotenv>=1.0.0

# Optional - for CLI and testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
